Add CommonMark auto download feature
This tool is dependent on the CommonMark provided test suite. Up until
now, it is the responsibility of the user to download the test suite.
This feature allows the tool to automatically download the test suite if
the user does not specify a path to the test suite. The tool will always
download the latest known tag at the time of publishing of tool.

- Reorganize Node.js dependencies to a native and external groups in
  cli.bin.js
- Add latestKnownTag property to app constant in cli.bin.js
- Add __spawnChildProcess() to cli.bin.js
- Add __gitVersion() to cli.bin.js
- Add __gitCheckoutCommonMarkLatest() to cli.bin.js
- Add downloadCommonMark() to cli.bin.js
- Reorganize cli entries in the arguments exclusion list in cli.bin.js
- Add commonmark-test-viewer entries to the arguments exclusion list in
  cli.bin.js
- Add ctv entries to the arguments exclusion list in cli.bin.js
- Reorganize banner() call to mark the end of the initialization phase
  of the tool's bootstrap in cli.bin.js
- Add a call to downloadCommonMark() if spec_tests.py is not provided to
  cli.bin.js